+++ app/src/main/java/com/ardakazanci/weatherappv2/Model/Wind.java
@@ -3,7 +3,7 @@ package com.ardakazanci.weatherappv2.Model;
 public class Wind {
 
     private double speed;
-    private int deg;
+    private double deg;
 
     public double getSpeed() {
         return speed;
@@ -13,7 +13,7 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public double getDeg() {
         return deg;
     }
